requirements.txt
netmiko
datetime
ipaddress
sys
time
urllib3
pytz
requests
pyserial
timezonefinder
tqdm
